+++ internal/usecase/repo/table_name.go
@@ -1,0 +1,8 @@
+package repo
+
+const (
+	tableNameStock    string = "basic_stock"
+	tableNameCalendar string = "basic_calendar"
+	tableNameEvent    string = "sinopac_event"
+	tableNameTarget   string = "basic_targets"
+)