+++ internal/usecase/repo/target_postgres.go
@@ -2,13 +2,10 @@ package repo
 
 import (
 	"context"
+	"time"
 
 	"toc-machine-trading/internal/entity"
 	"toc-machine-trading/pkg/postgres"
-)
-
-const (
-	tableNameTarget string = "basic_targets"
 )
 
 // TargetRepo -.
@@ -28,27 +25,26 @@ func (r *TargetRepo) InsertTargetArr(ctx context.Context, t []*entity.Target) er
 		builder = builder.Values(v.StockNum, v.TradeDay, v.Rank, v.Volume, v.Subscribe, v.RealTimeAdd)
 	}
 
-	sql, args, err := builder.ToSql()
-	if err != nil {
+	if sql, args, err := builder.ToSql(); err != nil {
+		return err
+	} else if _, err := r.Pool.Exec(ctx, sql, args...); err != nil {
 		return err
 	}
-	_, err = r.Pool.Exec(ctx, sql, args...)
-	if err != nil {
-		return err
-	}
-
 	return nil
 }
 
-// QueryAllTargetByTradeDay -.
-func (r *TargetRepo) QueryAllTargetByTradeDay(ctx context.Context) ([]*entity.Target, error) {
-	sql, _, err := r.Builder.
-		Select("*").From(tableNameTarget).ToSql()
+// QueryTargetsByTradeDay -.
+func (r *TargetRepo) QueryTargetsByTradeDay(ctx context.Context, tradeDay time.Time) ([]*entity.Target, error) {
+	sql, args, err := r.Builder.
+		Select("id, rank, volume, subscribe, real_time_add, trade_day, stock_num, number, name, exchange, category, day_trade, last_close").
+		From(tableNameTarget).
+		Where("trade_day = ?", tradeDay).
+		Join("basic_stock ON basic_targets.stock_num = basic_stock.number").ToSql()
 	if err != nil {
 		return nil, err
 	}
 
-	rows, err := r.Pool.Query(ctx, sql)
+	rows, err := r.Pool.Query(ctx, sql, args...)
 	if err != nil {
 		return nil, err
 	}
@@ -56,14 +52,14 @@ func (r *TargetRepo) QueryAllTargetByTradeDay(ctx context.Context) ([]*entity.Ta
 
 	entities := make([]*entity.Target, 0, 256)
 	for rows.Next() {
-		e := &entity.Target{}
-		err = rows.Scan(
-			&e.StockNum, &e.TradeDay, &e.Rank, &e.Volume, &e.Subscribe, &e.RealTimeAdd,
-		)
-		if err != nil {
+		e := entity.Target{Stock: new(entity.Stock)}
+		if err := rows.Scan(
+			&e.ID, &e.Rank, &e.Volume, &e.Subscribe, &e.RealTimeAdd, &e.TradeDay,
+			&e.StockNum, &e.Stock.Number, &e.Stock.Name, &e.Stock.Exchange, &e.Stock.Category, &e.Stock.DayTrade, &e.Stock.LastClose,
+		); err != nil {
 			return nil, err
 		}
-		entities = append(entities, e)
+		entities = append(entities, &e)
 	}
 	return entities, nil
 }