+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# (2022-06-06)
+# (2022-06-07)
 
 ### Bug Fixes
 
@@ -13,4 +13,6 @@
 * **event:** add eventbus package, rename from stock to basic ([a66f229](https://gitlab.tocraw.com/root/toc-machine-trading/commit/a66f229ce8fd5a137a613bd34c6485ad80cc7068))
 * **logger:** add caller func ([94fd4a5](https://gitlab.tocraw.com/root/toc-machine-trading/commit/94fd4a56259d9ca8e7c99417195165358148b2f8))
 * **naming:** make clean arch naming make sense ([538eb30](https://gitlab.tocraw.com/root/toc-machine-trading/commit/538eb3051f2719f90b288ef004644591be4b0406))
+* **sinopac:** implement all sinopac gRPC method ([2b868a9](https://gitlab.tocraw.com/root/toc-machine-trading/commit/2b868a96da2aee5f22c0fd4a2c8f974fcc071159))
+* **stream:** add stream usecase first part, move pb package to outside ([240b8ee](https://gitlab.tocraw.com/root/toc-machine-trading/commit/240b8ee886cb76486e78735d2cbf4f815fba9f8f))
 * **usecase:** add first usecase, include api, grpc, postgres repo ([9800f86](https://gitlab.tocraw.com/root/toc-machine-trading/commit/9800f865e3ed0b16afe06f487523c0a3a4d2b9d8))